refeat: 폴더 및 파일 정리
각 용도에 맞는 파일 정리
