+++ Calculator/Calculator/View/Assets.xcassets/Contents.json
